+++ R/geom_gene_arrow.R
@@ -108,7 +108,7 @@ GeomGeneArrow <- ggplot2::ggproto(
       width = grid::unit(1, "npc") - grid::unit(1, "mm"),
       height = grid::unit(1, "npc") - grid::unit(1, "mm"),
       gp = grid::gpar(
-        col = data$colour,
+        col = ggplot2::alpha(data$colour, data$alpha),
         fill = ggplot2::alpha(data$fill, data$alpha),
         lty = data$linetype,
         # linewidth is expressed in mm but grid expects points; multiplying by